Registration: Valid Until Should be INT (#6370)
* Registration: Valid Until Should be INT

See: https://mantis.ilias.de/view.php?id=38057
(cherry picked from commit 56b3ec38ae4130df8125e575595bf9c1ecfb0230)

+++ Services/Registration/classes/class.ilAccountRegistrationGUI.php
@@ -446,7 +446,7 @@ class ilAccountRegistrationGUI
 
                         case "relative":
                             $rel = unserialize($code_data["alimitdt"], ['allowed_classes' => false]);
-                            $access_limit = $rel["d"] * 86400 + $rel["m"] * 2592000 + $rel["y"] * 31536000 + time();
+                            $access_limit = (int) ($rel["d"] * 86400 + $rel["m"] * 2592000 + $rel["y"] * 31536000 + time());
                             break;
                     }
                 }
